fix(Builder) too early is_blog_installed run
This changes the logic put in place in [this
commit](https://github.com/stellarwp/schema/commit/fba6dc43e364b0212279ba13b6a2ee6fa1b09c75)
to keep the spirit (avoid a db query on each run of the `Builder::up()`
method, but avoid the call to `is_blog_installed` that, when fired too
early in multisite installations, would kill the database.

In production this would happen only once, but in tests it would crash
the multisite suites consistently when the table registration is
triggered.

I've verified the fix in the environment where I was getting the issue,
the `muwpunit` suite of TEC/Common.

The theory behind the fix is that the `is_blog_installed()` function
runs a very opinionated check that involves killing the database if the
database schema is not correct (e.g. in test there would be a `users`
table, but not an `option` one). Internally the function will set the
`is_blog_installed` cached value reliably when triggered in its natural
order of operations.

Why would, then, the `is_blog_installed()` function called too early?
Registering one table in a multi-site context would trigger the call to
`Builder::update_blog_tables` that would fire when the blog is changed
to install the new blog. The `users` table, common to all blogs, would
be there, but the `2_options` table of the blog we're switching to (e.g.
with id `2`) would not. So `is_blog_installed()` would `die` with a dead
db message.
So the blog **is being installed**, it's not installed.

With this change custom tables are not being created during a blog
installation in multisite context, is it fine? Yes, the next request for
the blog, or the next request switching to the blog, will find the blog
installed.