fix(Builder) too early is_blog_installed run
This changes the logic put in place in [this
commit](https://github.com/stellarwp/schema/commit/fba6dc43e364b0212279ba13b6a2ee6fa1b09c75)
to keep the spirit (avoid a db query on each run of the `Builder::up()`
method, but avoid the call to `is_blog_installed` that, when fired too
early in multisite installations, would kill the database.

What is the issue?
The `Builder::up` is called, by means of the
`Builder::update_blog_tables` on `switch_blog`.
The `switch_blog` function is called during the creation of blogs as
well, in the `wp_initialize_site` function.
The blog **is being installed**, it's not installed yet.
Running `is_blog_installed()` in this phase will kill the site as it
will find a broken db: the `users` table is there (it' common to all
blogs), but the `options` table is not.

This PR handles this scenario, delaying the creation of the tables to
the `activate_blog` or `switch_blog` calls following the creation of the
site or future requests.